backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymysql==1.1.2
Werkzeug==3.0.1
python-dotenv==1.0.0
pandas>=1.5.0
scikit-learn>=1.0.0
numpy>=1.20.0
shapely>=2.0.0
